Created functions to add and delete reactions to thoughts.

+++ controllers/userController.js
@@ -34,58 +34,63 @@ function updateUser(req, res) {
 
 // GET a single user
 function getOneUser(req, res) {
-  User.findOne(
-    { _id: req.params.user_id } 
-  )
-  .then((user) => {
-    if(!user) {
-      return res.status(404).json({ message: 'No user with this id'})
-    }
-    res.json(user)
-  })
-  .catch((err) => res.status(500).json(err));
+  User.findOne({ _id: req.params.user_id })
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({ message: "No user with this id" });
+      }
+      res.json(user);
+    })
+    .catch((err) => res.status(500).json(err));
 }
 
-// DELETE a user 
-function deleteUser(req,res) {
-  User.findOneAndDelete(
-    { _id: req.params.user_id }
-  )
-  .then((user) => res.json(user))
-  .catch((err) => res.status(500).json(err));
+// DELETE a user
+function deleteUser(req, res) {
+  User.findOneAndDelete({ _id: req.params.user_id })
+    .then((user) => res.json(user))
+    .catch((err) => res.status(500).json(err));
 }
 
+// ----------------------------------------------------------------------------------------------------------------------
 // POST a friend to a user
-function addFriend(req,res) {
+function addFriend(req, res) {
   User.findOneAndUpdate(
     { _id: req.params.user_id },
-    { $addToSet: {friends: req.params.friends_id}},
+    { $addToSet: { friends: req.params.friends_id } },
     { new: true }
   )
-  .then((user) => {
-    if(!user) {
-      return res.status(404).json({ message: 'No user with this id'})
-    }
-    res.json(user)
-  })
-  .catch((err) => res.status(500).json(err))
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({ message: "No user with this id" });
+      }
+      res.json(user);
+    })
+    .catch((err) => res.status(500).json(err));
 }
 
 // DELETE a friend from a user
-function deleteFriend(req,res) {
+function deleteFriend(req, res) {
   User.findOneAndUpdate(
     { _id: req.params.user_id },
-    { $pull: { friends: req.params.friends_id}},
-    { new: true}
+    { $pull: { friends: req.params.friends_id } },
+    { new: true }
   )
-  .then((user) => {
-    if(!user) {
-      return res.status(404).json({ message: 'No user with this id'})
-    }
-    res.json(user)
-  })
-  .catch((err) => res.status(500).json(err))
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({ message: "No user with this id" });
+      }
+      res.json(user);
+    })
+    .catch((err) => res.status(500).json(err));
 }
 
 // Exporting functions out
-module.exports = { createUser, getUsers, updateUser, getOneUser, deleteUser, addFriend, deleteFriend };
+module.exports = {
+  createUser,
+  getUsers,
+  updateUser,
+  getOneUser,
+  deleteUser,
+  addFriend,
+  deleteFriend,
+};
